case_management_service/requirements.txt
confluent-kafka
pydantic
pymongo
fastapi
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-logging
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-pymongo
python-json-logger
